Move test_employee_register and test_list_employee to test_employee
